+++ include/alcelin_file_utilities.hpp
@@ -39,10 +39,12 @@
 
 #pragma once
 
+#include <algorithm>
 #include <cstring>
 #include <fstream>
 #include <ios>
 #include <stdexcept>
+#include <vector>
 
 /**
  *  @brief  All Alcelin's contents in this namespace.  Just do
@@ -54,21 +56,26 @@ namespace alcelin {
  *  @brief  File utilities.
  *
  *  Contains utilities regarding file handling, and functionality to handle
- *  binary files of SD format.
+ *  binary file data in SD format.
  *
  *  SD Format is a binary format which consists of chunks.  Each chunk is a pair
- *  of *Size* and *Data*.  The Size is the number of bytes of the raw data.  The
- *  Data is the raw data, i.e., the memory map of an object.
+ *  of *Size* and *Data*.  The Size part of the data is @c std::size_t bytes and
+ *  specifies the size of the chunk to be read to get the entire chunk.
  *
  *  @note  The SD Format does not specify the endianness of the data, neither
  *         does the code to handle this format.  This implies that the binary
  *         file may not be portable for systems with different endianness.
  *  @todo  When C++ introduces a good reflection support, implement automatic
  *         endianness conversion to make the binary file actually be portable.
- *
- *  @todo  The biggest todo: implement stuff.
+ *         but for now, this is an impossible task and no plans to support it
+ *         with hack codes.
  */
 namespace file {
+
+/**
+ *  @brief  A "SD chunk" is basically a vector of bytes.
+ */
+using sd_chunk = std::vector<std::byte>;
 
 /**
  *  @brief  Read all the file's content at once and return @c std::string
@@ -92,62 +99,113 @@ namespace file {
 }
 
 /**
- *  @brief  A chunk in the SD file format.
- *  @see  Detailed Description of namespace @c file.
- */
-struct sd_chunk {
-
-    /**
-     *  @brief  The number of bytes of @c data .
-     */
-    std::size_t size;
-
-    /**
-     *  @brief  The raw data of @c size bytes.
-     *
-     *  @note  The values pointed by the pointer are not changed.  It is user's
-     *         responsibility to manage the memory.
-     *  @todo  If nobody finds a use for `void *` data type for pointer for
-     *         chunks or finds it tedious to manage memory (like we do), replace
-     *         it with @c std::shared_ptr<void> .
-     */
-    void *data;
-};
-
-/**
- *  @brief  Convert an object to a chunk.
+ *  @brief  Creates a chunk from object @c t .
  *
  *  @tparam  type  Trivially copyable type.
  *  @param   t     Trivially copyable value.
- *  @return  A newly allocated chunk.  Memory allocation failure may occur.
- *
- *  @note  Use @c delete_sd_chunk to free the allocated memory.
+ *  @return  Chunk of size `sizeof (type)` and containing the data from @c t .
  */
 template<typename type>
 requires(std::is_trivially_copyable_v<type>)
-[[nodiscard]] inline constexpr auto to_new_sd_chunk(const type &t) -> sd_chunk;
+[[nodiscard]] inline constexpr auto to_sd_chunk(const type &t)
+{
+    auto     t_size = sizeof (type);
+    sd_chunk chunk(t_size);
+
+    std::memcpy(chunk.data(), &t, t_size);
+    return chunk;
+}
 
 /**
- *  @brief  Convert a chunk to an object.
+ *  @brief  Convert a @c type from chunk.
  *
  *  @tparam  type   Trivially copyable chunk.
- *  @param   chunk  Chunk with memory layout of type @c T .
- *  @return  An object of T with memory layout of the chunk.
+ *  @param   chunk  Chunk with size and data of @c type .
+ *  @return  Object of type @c type with data of the chunk.
  */
 template<typename type>
 requires(std::is_trivially_copyable_v<type>)
-[[nodiscard]] inline constexpr auto from_sd_chunk(sd_chunk chunk) -> type;
+[[nodiscard]] inline constexpr auto from_sd_chunk(const sd_chunk &chunk)
+{
+    auto t_size = sizeof (type);
+
+    if (t_size != chunk.size())
+    {
+        throw std::invalid_argument(std::format(
+            "SD chunk size ({}) does not match type size ({})", chunk.size(),
+            t_size));
+    }
+
+    auto t = type();
+    std::memcpy(&t, chunk.data(), t_size);
+    return t;
+}
 
 /**
- *  @brief  Delete the memory of a chunk.
+ *  @brief  Read a chunk from an input stream.
  *
- *  After creating a chunk using @c to_new_sd_chunk , using it and being done
- *  with it, be sure to call this function to delete the memory of the created
- *  chunk.
- *
- *  @param  chunk  Chunk to delete.
+ *  @param  input  Input stream to read from.
+ *  @return  Chunk read from the stream.
  */
-inline constexpr auto delete_sd_chunk(sd_chunk chunk) -> void;
+[[nodiscard]] inline constexpr auto read_chunk(std::istream &input)
+{
+    std::size_t size = 0;
+    input.read((char *)&size, sizeof(std::size_t));
+
+    sd_chunk chunk(size);
+    input.read((char *)chunk.data(), chunk.size());
+    return chunk;
+}
+
+/**
+ *  @brief  Write a chunk to an output stream.
+ *
+ *  @param  output  Output stream to write to.
+ *  @param  chunk   Chunk to write.
+ */
+[[nodiscard]] inline constexpr auto write_chunk(
+    std::ostream &output,
+    const sd_chunk &chunk
+)
+{
+    std::size_t size = chunk.size();
+    output.write((char *)&size, sizeof(std::size_t));
+
+    output.write((char *)chunk.data(), chunk.size());
+}
+
+/**
+ *  @brief  Helper to read a chunk and convert it to type.
+ *
+ *  @tparam  type    Trivially copyable type.
+ *  @param   input   Input stream to read from.
+ *  @return  Data converted from chunk read from the stream.
+ */
+template<typename type>
+requires(std::is_trivially_copyable_v<type>)
+[[nodiscard]] inline constexpr auto read_data(std::istream &input)
+{
+    auto chunk = read_chunk(input);
+    return from_sd_chunk<type>(chunk);
+}
+
+/**
+ *  @brief  Helper to convert the type to chunk and write it.
+ *  
+ *  @tparam  type    Trivially copyable type.
+ *  @param   output  Output stream to write to.
+ *  @param   t       Trivially copyable value.
+ */
+template<typename type>
+requires(std::is_trivially_copyable_v<type>)
+[[nodiscard]] inline constexpr auto write_data(
+    std::ostream &output,
+    const type &t
+)
+{
+    auto chunk = to_sd_chunk(t);
+    write_chunk(output, chunk);
+}
 
 } // namespace file